Extract marker overflow dialog helpers
